requirements.txt
altgraph==0.17.4
BrainLinkParser==0.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
cushy-serial==1.1.1
idna==3.10
packaging==24.2
pefile==2023.2.7
pip==25.0
pyasn1==0.6.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyserial==3.5
PySimpleGUI==5.0.9
pywin32-ctypes==0.2.3
requests==2.32.3
rsa==4.9
setuptools==75.8.0
urllib3==2.3.0
wheel==0.45.1